completed refactor for editor directory
