+++ lua/plugins/editor/symbols-outline/keys.lua
@@ -1,0 +1,3 @@
+return {
+	{ "<leader>so", "<cmd>SymbolsOutline<cr>", desc = "Outline symbols in a file" },
+}
