+++ lua/plugins/editor/symbols-outline/opts.lua
