+++ lua/plugins/editor/todo-comments/keys.lua
@@ -1,0 +1,3 @@
+return {
+	{ "<leader>tc", "<cmd>TodoTelescope<cr>", desc = "Search for TODOs" },
+}
