+++ lua/plugins/editor/vim-maximizer/keys.lua
@@ -1,0 +1,3 @@
+return {
+	{ "<leader>sm", "<cmd>MaximizerToggle<cr>", desc = "Toggle between maximze pane and normal pane size" },
+}
